+++ App/Source/GameComponents/GameLayer.cpp
@@ -7,55 +7,55 @@
 #include <stdexcept>
 #include <array>
 
-Vector2Ex<int> GameLayer::GetNextValidPacmanPosition(const UIComponents::Direction direction) const
+Vector2Ex<float> GameLayer::GetNextValidPacmanPosition(const UIComponents::Direction direction) const
 {
     using namespace UIComponents;
     using enum Direction;
 
-    const Vector2Ex<int> currentPosition = m_pacman.GetPositionAtAnchor();
-    const Vector2Ex<int> nextPosition = m_pacman.GetNextPosition(direction);
+    const Vector2Ex<float> currentPosition = m_pacman.GetPositionAtAnchor();
+    const Vector2Ex<float> nextPosition = m_pacman.GetNextPosition(direction);
 
-    const Vector2Ex<int> movementDelta = nextPosition - currentPosition;
+    const Vector2Ex<float> movementDelta = nextPosition - currentPosition;
 
-    int steps;
+    float steps;
 
     if (direction == UP || direction == DOWN)
         steps = std::abs(movementDelta.y);
     else
         steps = std::abs(movementDelta.x);
 
-    Vector2Ex<int> nextValidPosition = currentPosition;
-    for (int step = 0; step <= steps; step++)
+    Vector2Ex<float> nextValidPosition = currentPosition;
+    for (float step = 0; step <= steps; step++)
     {
-        Vector2Ex<int> intermediatePosition;
+        Vector2Ex<float> intermediatePosition;
         if (steps == 0)
             intermediatePosition = currentPosition;
         else
             intermediatePosition = currentPosition + (movementDelta / steps) * step;
 
-        Vector2Ex<int> cornersToCheck[2];
-        const Vector2Ex<int> pacmanDimensions = m_pacman.GetDimensions();
+        Vector2Ex<float> cornersToCheck[2];
+        const Vector2Ex<float> pacmanDimensions = m_pacman.GetDimensions();
 
         switch (direction)
         {
         case UP:
             cornersToCheck[0] = intermediatePosition;
-            cornersToCheck[1] = intermediatePosition + Vector2Ex<int>(pacmanDimensions.x, 0);
+            cornersToCheck[1] = intermediatePosition + Vector2Ex<float>(pacmanDimensions.x, 0);
             break;
 
         case DOWN:
-            cornersToCheck[0] = intermediatePosition + Vector2Ex<int>(0, pacmanDimensions.y);
-            cornersToCheck[1] = intermediatePosition + Vector2Ex<int>(pacmanDimensions.x, pacmanDimensions.y);
+            cornersToCheck[0] = intermediatePosition + Vector2Ex<float>(0, pacmanDimensions.y);
+            cornersToCheck[1] = intermediatePosition + Vector2Ex<float>(pacmanDimensions.x, pacmanDimensions.y);
             break;
 
         case LEFT:
             cornersToCheck[0] = intermediatePosition;
-            cornersToCheck[1] = intermediatePosition + Vector2Ex<int>(0, pacmanDimensions.y);
+            cornersToCheck[1] = intermediatePosition + Vector2Ex<float>(0, pacmanDimensions.y);
             break;
 
         case RIGHT:
-            cornersToCheck[0] = intermediatePosition + Vector2Ex<int>(pacmanDimensions.x, 0);
-            cornersToCheck[1] = intermediatePosition + Vector2Ex<int>(pacmanDimensions.x, pacmanDimensions.y);
+            cornersToCheck[0] = intermediatePosition + Vector2Ex<float>(pacmanDimensions.x, 0);
+            cornersToCheck[1] = intermediatePosition + Vector2Ex<float>(pacmanDimensions.x, pacmanDimensions.y);
             break;
         }
 
@@ -87,7 +87,7 @@ Vector2Ex<int> GameLayer::GetNextValidPacmanPosition(const UIComponents::Directi
 
 GameLayer::GameLayer()
     : m_board(Board()),
-      m_pacman(m_board.GetPositionFromIndex(Vector2Ex<int>(1, 1)) + Vector2Ex<int>(1, 1), Vector2Ex<int>(48, 48), 5)
+      m_pacman(m_board.GetPositionFromIndex(Vector2Ex<int>(1, 1)) + Vector2Ex<float>(1, 1), Vector2Ex<float>(48, 48), 5)
 {
 }
 
@@ -110,8 +110,8 @@ void GameLayer::HandleKeyPresses()
 void GameLayer::HandleCollisions()
 {
 
-    const Vector2Ex<int> currentPosition = m_pacman.GetPositionAtAnchor();
-    const Vector2Ex<int> nextQueuedPosition = GetNextValidPacmanPosition(m_pacman.GetQueuedDirection());
+    const Vector2Ex<float> currentPosition = m_pacman.GetPositionAtAnchor();
+    const Vector2Ex<float> nextQueuedPosition = GetNextValidPacmanPosition(m_pacman.GetQueuedDirection());
 
     if (currentPosition != nextQueuedPosition)
         m_pacman.ApplyQueuedDirection();
