+++ Core/Source/DataTypes/Bounds.h
@@ -5,13 +5,13 @@
 class Bounds
 {
 public:
-    Vector2Ex<int> dimensions;
-    Vector2Ex<int> position; // Top left (drawpoint)
+    Vector2Ex<float> dimensions;
+    Vector2Ex<float> position; // Top left (drawpoint)
 
 public:
-    Bounds(Vector2Ex<int> dimensions, Vector2Ex<int> position);
+    Bounds(Vector2Ex<float> dimensions, Vector2Ex<float> position);
     ~Bounds() = default;
 
-    bool IsPointInBounds(const Vector2Ex<int> &position) const;
+    bool IsPointInBounds(const Vector2Ex<float> &position) const;
     static bool IsBoundsTouching(const Bounds &a, const Bounds &b);
 };
