+++ Core/Source/UIComponents/Text.h
@@ -18,14 +18,14 @@ namespace UIComponents
         TextStyle m_style;
 
     public:
-        Text(std::string text, TextStyle style, Vector2Ex<int> anchorPointPosition, AnchorPoint AnchorPoint, bool visible);
+        Text(std::string text, TextStyle style, Vector2Ex<float> anchorPointPosition, AnchorPoint AnchorPoint, bool visible);
         ~Text() = default;
 
         void SetText(std::string text);
         std::string GetText() const;
 
-        void SetFontSize(int fontSize);
-        int GetFontSize() const;
+        void SetFontSize(float fontSize);
+        float GetFontSize() const;
 
         void SetColor(Color color);
         Color GetColor() const;
@@ -33,8 +33,8 @@ namespace UIComponents
         void SetStyle(TextStyle style);
         TextStyle GetStyle() const;
 
-        virtual Vector2Ex<int> GetDimensions() const override;
-        virtual void Render(Vector2Ex<int> offset = {0, 0}) const override;
+        virtual Vector2Ex<float> GetDimensions() const override;
+        virtual void Render(Vector2Ex<float> offset = {0, 0}) const override;
     };
 
 }
